feat(architecture): align route composition and remove legacy paths
