--- src/data/blogPost.ts
@@ -1,50 +1,0 @@
-export type BlogCard = {
-  id: number;
-  category: string;
-  title: string;
-  imageUrl: string;
-  authorName: string;
-  authorAvatar: string;
-  description: string;
-  date: string;
-};
-
-const blogPosts: BlogCard[] = [
-  {
-    id: 1,
-    category: "Tech & Code",
-    title: "Why Clean Architecture is a Lifesaver (Trust Me, I’ve Been There)",
-    imageUrl: "/images/dummy-image.jpg",
-    authorName: "You",
-    authorAvatar: "/images/avatar.png",
-    description:
-      "Ever had a project turn into spaghetti code faster than instant noodles? Let’s talk about why Clean Architecture saved me (and my sanity).",
-    date: "2025-01-01",
-  },
-  {
-    id: 2,
-    category: "Dev Life",
-    title:
-      "Handling a Flutter Project Alone: Fun, Pain, and Late-Night Debugging",
-    imageUrl: "/images/dummy-image.jpg",
-    authorName: "You",
-    authorAvatar: "/images/avatar.png",
-    description:
-      "Took on a full Flutter project solo? Here’s what I learned (including how not to lose your mind in the process).",
-    date: "2025-01-05",
-  },
-  {
-    id: 3,
-    category: "Real Talk",
-    title: "The Time I Tried to Do Everything Myself (And Why You Shouldn’t)",
-    imageUrl: "/images/dummy-image.jpg",
-    authorName: "You",
-    authorAvatar: "/images/avatar.png",
-    description:
-      "Spoiler: It didn’t end well. Here’s a raw take on why teamwork actually matters and how I learned the hard way.",
-    date: "2025-01-10",
-  },
-];
-
-export default blogPosts;
-
