--- src/data/projects.ts
@@ -1,45 +1,0 @@
-export type ProjectCard = {
-  title: string;
-  description: string;
-  link: string;
-  borderStyles: string;
-  image: string;
-};
-
-const projects: ProjectCard[] = [
-  {
-    title: "eDamkar",
-    description:
-      "Lorem ipsum dolor sit amet, Lorem ipsum dolor sit amet, Lorem ipsum dolor sit amet.",
-    link: "/projects/edamkar",
-    borderStyles: "border-b lg:border-b border-gray-200",
-    image: "/images/dummy-image.jpg",
-  },
-  {
-    title: "ePKK",
-    description:
-      "Lorem ipsum dolor sit amet, Lorem ipsum dolor sit amet, Lorem ipsum dolor sit amet.",
-    link: "/projects/epkk",
-    borderStyles: "border-b lg:border-b lg:border-l border-gray-200",
-    image: "/images/dummy-image.jpg",
-  },
-  {
-    title: "Ngawi Smart City",
-    description:
-      "Lorem ipsum dolor sit amet, Lorem ipsum dolor sit amet, Lorem ipsum dolor sit amet.",
-    link: "/projects/ngawi-smart-city",
-    borderStyles: "border-b lg:border-none border-gray-200",
-    image: "/images/dummy-image.jpg",
-  },
-  {
-    title: "Lofo - leftover food",
-    description:
-      "Lorem ipsum dolor sit amet, Lorem ipsum dolor sit amet, Lorem ipsum dolor sit amet.",
-    link: "/projects/lofo",
-    borderStyles: "lg:border-l border-gray-200",
-    image: "/images/dummy-image.jpg",
-  },
-];
-
-export default projects;
-
