--- src/data/skill.ts
@@ -1,25 +1,0 @@
-export type SkillLink = {
-  name: string;
-  link: string;
-  icon: string;
-};
-
-const skills: SkillLink[] = [
-  {
-    name: "LinkedIn",
-    link: "https://www.linkedin.com/in/fikrilal/",
-    icon: "/icons/icons8-linkedin.svg",
-  },
-  {
-    name: "Instagram",
-    link: "https://www.instagram.com/fikril.al/",
-    icon: "/icons/icons8-instagram.svg",
-  },
-  {
-    name: "Email",
-    link: "mailto:fikrildev@gmail.com",
-    icon: "/icons/icons8-x.svg",
-  },
-];
-
-export default skills;
