+++ src/features/work/components/WorkIndexPage.tsx
@@ -1,0 +1,24 @@
+import Footer from "@/components/Footer";
+import Navbar from "@/components/Navbar";
+import SelectedWork from "@/components/SelectedWork";
+import { getWorkSummaries } from "@/lib/work";
+
+export default async function WorkIndexPage() {
+  const workItems = await getWorkSummaries();
+
+  return (
+    <>
+      <Navbar />
+      <main className="bg-neutral-950">
+        <SelectedWork
+          workItems={workItems}
+          heading="All Work & Case Studies"
+          description="A deeper look at the products and case studies I’ve delivered—from mobile apps to full-stack builds."
+          ctaHref=""
+          ctaLabel=""
+        />
+      </main>
+      <Footer />
+    </>
+  );
+}
